// src/utils/Enums.ts
export enum MapName {
    CLASSIC = 'Classic',
    THE_GREAT_DIVIDE = 'The Great Divide',
    ARCHIPELAGO_ASSAULT = 'Archipelago Assault',
    PACMAN = 'Pacman',
}

export enum Orientation {
    HORIZONTAL = "Horizontal",
    VERTICAL = "Vertical",
}

export enum ShipName {
    SUBMARINE = "Submarine",
    SUPPLY_BOAT = "Supply Boat",
    DESTROYER = "Destroyer",
    BATTLESHIP = "Battleship",
    FRIGATE = "Frigate",
    AIRCRAFT_CARRIER = "Aircraft Carrier",
}